// apps/api/src/generated/ticktick/types.gen.ts
export type ChecklistItem = {
  /**
   * The ID of checklistItem
   */
  id?: string
  /**
   * Title of checklistItem
   */
  title?: string
  /**
   * The completion status of checklistItem, Normal is 0, Completed is 1
   */
  status?: number
  /**
   * checklistItem completed time in "yyyy-MM-dd'T'HH:mm:ssZ", Example "2019-11-13T03:00:00+0000
   */
  completedTime?: string
  /**
   * Is checklistItem all day, true or false
   */
  isAllDay?: string
  /**
   * Order value of checklistItem
   */
  sortOrder?: number
  /**
   * checklistItem start date time in "yyyy-MM-dd'T'HH:mm:ssZ", Example "2019-11-13T03:00:00+0000"
   */
  startDate?: string
  /**
   * checklistItem timezone, Example "America/Los_Angeles"
   */
  timeZone?: string
}

export type Task = {
  /**
   * The ID of task
   */
  id?: string
  /**
   * The ID of project which task in
   */
  projectId?: string
  /**
   * Title of task
   */
  title?: string
  /**
   * Is task all day, true or false
   */
  isAllDay?: string
  /**
   * Task completed time in "yyyy-MM-dd'T'HH:mm:ssZ", Example "2019-11-13T03:00:00+0000
   */
  completedTime?: string
  /**
   * Content of task
   */
  content?: string
  /**
   * Description of checklist
   */
  desc?: string
  /**
   * Task due date time in "yyyy-MM-dd'T'HH:mm:ssZ", Example "2019-11-13T03:00:00+0000"
   */
  dueDate?: string
  items?: Array<ChecklistItem>
  /**
   * Task priority, None is 0, Low is 1, Medium is 3, High is 5
   */
  priority?: number
  /**
   * List of reminder trigger, Example ["TRIGGER:P0DT9H0M0S","TRIGGER:PT0S"]
   */
  reminders?: Array<string>
  /**
   * Recurring rules of task, Example "RRULE:FREQ=DAILY;INTERVAL=1"
   */
  repeatFlag?: string
  /**
   * Sort order value of task, Example 12345
   */
  sortOrder?: number
  /**
   * Start data time in "yyyy-MM-dd'T'HH:mm:ssZ", Example "2023-04-23T12:00:00+0000"
   */
  startDate?: string
  /**
   * Task completion status, Normal is 0, Completed is 2
   */
  status?: number
  /**
   * Task timezone, Example "America/Los_Angeles"
   */
  timeZone?: string
}

export type Project = {
  /**
   * The ID of project
   */
  id?: string
  /**
   * Name of project
   */
  name?: string
  /**
   * Color of project
   */
  color?: string
  /**
   * Order value of project
   */
  sortOrder?: number
  /**
   * Project closed, true or false
   */
  closed?: string
  /**
   * ID of project group
   */
  groupId?: string
  /**
   * View mode of project, "list" or "kanban" or "timeline"
   */
  viewMode?: string
  /**
   * Permission of project, "read" or "write" or "comment"
   */
  permission?: string
  /**
   * Kind of project, "Task" or "NOTE"
   */
  kind?: string
}